API解析を待機するように、ほか
API解析の待機
画面サイズの変更
終了ダイアログの変更

+++ KanColleLib/TransmissionData/api_get_member/Questlist.cs
@@ -46,11 +46,12 @@ namespace KanColleLib.TransmissionData.api_get_member
                 page_count = (int)json.api_page_count,
                 disp_page = (int)json.api_disp_page,
                 exec_count = (int)json.api_exec_count,
-                exec_type = (int)json.api_exec_type
+                exec_type = (long)json.api_exec_type
             };
 
             questlist.list = new List<values.QuestlistValue>();
 
+            // svdata={"api_result":1,"api_result_msg":"\u6210\u529f","api_data":{"api_count":20,"api_page_count":4,"api_disp_page":5,"api_list":null,"api_exec_count":4,"api_exec_type":2357015}}
             if (json.api_list != null)
             {
                 foreach (var data in json.api_list)
